12.141 Routing - Vite exposes env variables on the special import.meta.env object. Some built-in variables are available in all cases: - import.meta.env.MODE: {string} the mode the app is running in.
- import.meta.env.BASE_URL: {string} the base url the app is being served from. This is determined by the base config option.

- import.meta.env.PROD: {boolean} whether the app is running in production.

- import.meta.env.DEV: {boolean} whether the app is running in development (always the opposite of import.meta.env.PROD)

- import.meta.env.SSR: {boolean} whether the app is running in the server.

Most of the latest browsers come with an API called the history API. It lets you interact with the browser history. You can use it to change the url in the address bar without having to refresh the page. The value for this must be an object with how we plan to use history API to navigate the app. There are a lot of things we need to account for when using the history API. Luckly the vue router library has a helper function for helping you generate the object. The value for this property is being set to the createWebHistory function. It just require to pass any base url. This will represent the base url for our application. The vue router library will take care of appending the url to the base url. This future saves us time for having to type out the full url constantly. We can focus on the paths. In addition it's important to set in case your application is inside a nested directory. Now our application is hosted in the root document. For this project we're using a property called import.meta.env.BASE_URL
import.meta object contains information on the curent module. Through this object we can access info on the current environment. That info can be accessed by looking inside the environment property. We are accessing the BASE_URL property. BASE_URL property can be configured through the vite configuration file. For details look here https://vitejs.dev/guide/env-and-mode.html. Our server can be accessed by navigating to localhost. We don't have to worry about adding additional directories.