13.162 Uploading files - Update the progress bar whenever the state change it
We don't know which object to update because there can be multiple objects in the array. We need to save the index after we insert it so that we know which object to update. We can save the index by assigning the value returned by the push function to a variable. Push function return the length of the array after the object has been pushed into it. We extract the index from it, because the object will be inserted as the last object in the array.